calibration of frequency and samplerate(s) for next hdsdr release (work in progress) (#143)
* calibration setting (in registry of hdsdr) modifies reported samplerates
 see https://github.com/ik1xpv/ExtIO_sddc/issues/110
 calibration settings has to be entered manually in the registry
 Computer\HKEY_CURRENT_USER\Software\HDSDR\ExtIO_SDDC.dll\012_Value
* calibration of hdsdr can be used to determine the ppm value ..
 but it must be resetted, when entering this value in the registry
 - whilst hdsdr is already terminated
* added new extio function 'ExtIoSrateSelText()'
 to display a readable samplerate in next hdsdr release
* added double variants for several extio functions
* deactivated unnecessary extio Tune functions

+++ libsddc/libsddc.cpp
@@ -276,7 +276,7 @@ double sddc_get_tuner_frequency(sddc_t *t)
 
 int sddc_set_tuner_frequency(sddc_t *t, double frequency)
 {
-    t->freq = t->handler->TuneLO(frequency);
+    t->freq = t->handler->TuneLO((int64_t)frequency);
 
     return 0;
 }
